[Update] 2리드 지원 + CPSC2021 + 라벨 맵핑 + README 전면 개편
- convert_to_h5: target_sig_name=None으로 원본 리드 그대로 저장 (2리드 등)
  CPSC2021 데이터셋 추가 (--group cpsc2021, prefix=c21)
- h5_structure: baseline int16→int32, adc_gain float16→float32 (overflow 수정)
- verify_h5: n_sig 가변 지원 (shape[0]!=12 하드코딩 제거)
- append_labels.py: 통합 라벨 맵핑 스크립트 신규
  데이터셋별 라벨 CSV/JSON을 각 H5 폴더에 생성
  지원: heedb(149), physionet(31), zzu(36), cpsc2021(3), code15(6)
- README: 전면 개편 — 데이터셋 현황, 라벨 현황, H5 템플릿 3종,
  출력 구조, 사용법, 권장 실행 순서

+++ utils/h5_structure.py
@@ -67,9 +67,9 @@ def create_h5_structure(
     if fmt is not None:
         meta_grp.create_dataset("fmt", data=np.array(fmt, dtype=UTF8), dtype=UTF8)
     if adc_gain is not None:
-        meta_grp.create_dataset("adc_gain", data=np.array(adc_gain, dtype=np.float16))
+        meta_grp.create_dataset("adc_gain", data=np.array(adc_gain, dtype=np.float32))
     if baseline is not None:
-        meta_grp.create_dataset("baseline", data=np.array(baseline, dtype=np.int16))
+        meta_grp.create_dataset("baseline", data=np.array(baseline, dtype=np.int32))
     if units is not None:
         meta_grp.create_dataset("units", data=np.array(units, dtype=UTF8), dtype=UTF8)
     if adc_res is not None:
